requirements.txt
anthropic>=0.83.0
mcp>=1.6.0
python-dotenv>=1.0.0
